+++ toDoList/toDoList/XYZAppDelegate.m
@@ -19,7 +19,7 @@
 // This is a global that can be accessed in any file that imports app delegate
 CLLocation *currentLoc;
 BOOL alertsOn = YES;
-double radiusScale = 1;
+float radiusScale = 1;
 
 - (void)locationManager:(CLLocationManager *)manager didUpdateToLocation:(CLLocation *)newLocation fromLocation:(CLLocation *)oldLocation
 {
